refactor(checker): split flow condition entrypoints
