+++ crates/tsz-checker/src/flow/control_flow/mod.rs
@@ -24,6 +24,7 @@ mod assignment_fallback;
 mod call_condition_narrowing;
 mod comparison_types;
 pub(crate) mod condition_narrowing;
+mod condition_narrowing_entrypoints;
 mod condition_nullish;
 mod core;
 mod flow_dp;
